FIR checker: utilize modifier retrievals
